requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bce-python-sdk==0.9.55
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
cryptography==46.0.3
dataclasses-json==0.6.7
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.121.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langsmith==0.4.42
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
numexpr==2.14.1
numpy==2.3.4
ollama==0.6.1
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
py-cpuinfo==9.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMuPDF==1.26.6
pypdfium2==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.6.0
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
